Debian Bug#157098: wrong file presentation on large files 	  from "Eduard Bloch" <blade@debian.org> * configure (clen_t): added 		(HAVE_STRTOLL): added 		(HAVE_STRTOQ): added 		(HAVE_ATOLL): added 		(HAVE_ATOQ): added * config.h.dist: ditto * file.c (current_content_length): s/int/clen_t/ 	(loadGeneralFile): s/atoi/strtoclen/ 	(convert_size): s/int/clen_t/ 	(convert_size2): s/int/clen_t/ 	(showProgress): s/int/clen_t/ 	(loadHTMLstream): s/int/clen_t/ linelen, trbyte 	(loadBuffer): ditto 	(getNextPage): s/int/clen_t/ linelen 	(save2tmp): s/int/clen_t/ linelen, trbye 	(_MoveFile): s/int/clen_t/ linelen, trbye * fm.h (_Buffer): s/int/clen_t/ linelen, trbye * ftp.c (size_int2str): s/long/clen_t/ 	(ex_ftpdir_name_size_date): s/long/clen_t/ * indep.c (strtoclen): added * indep.h (strtoclen): added * proto.h (showProgress): s/int/clen_t/ From: Fumitoshi UKAI  <ukai@debian.or.jp>

+++ fm.h
@@ -1,4 +1,4 @@
-/* $Id: fm.h,v 1.61 2002/07/22 16:27:31 ukai Exp $ */
+/* $Id: fm.h,v 1.62 2002/08/20 17:49:39 ukai Exp $ */
 /* 
  * w3m: WWW wo Miru utility
  * 
@@ -419,8 +419,8 @@ typedef struct _Buffer {
     struct frameset *frameset;
     struct frameset_queue *frameQ;
     int *clone;
-    int linelen;
-    int trbyte;
+    size_t linelen;
+    size_t trbyte;
     char check_url;
 #ifdef JP_CHARSET
     char document_code;
